ecs: Expect empty volume definition in --dry-run
When volume source_path is not given, ECS agent adds a emptyvolume container to
generate a nonpersistent empty volume then share between containers.

--dry-run could fail because it expects source_path. This patch changes
to give /tmp/ephemeral_$name as a source path instead when the path isn't given.

(To emulate the same behavior in command line requires creation of empty
image and container, but this patch doesn't implement because it is overkill)

https://github.com/aws/amazon-ecs-agent/blob/3631cb9de33a569d16e7a19d77105beeef8160c2/agent/api/task/task.go#L249
https://github.com/aws/amazon-ecs-agent/blob/3631cb9de33a569d16e7a19d77105beeef8160c2/agent/dockerclient/dockerapi/docker_client.go#L434
https://docs.aws.amazon.com/AmazonECS/latest/developerguide/using_data_volumes.html