Change threshold for detecting a deployment failure
Previously, hako fails a deployment when there's a task which is
launched by the deployment and stops during the deployment.  It wrongly
fails the deployment process when the container instance which a new
task is assigned to is changed to DRAINING status during the deployment.
So I decided to change the threshold as a mitigation. This is not a
complete solution though.