Load dependent containers on `hako oneshot`
Currently, `hako oneshot` loads the app container and sidecars linked
from the app container. However, this behavior is inconvenient when the
task contains an initializing sidecar. Suppose the app container uses a
Docker volume and the volume must be initialized first. This requirement
can be achieved as follows.

```jsonnet
{
  app: {
    depends_on: [
      { container_name: 'init', condition: 'SUCCESS' },
    ],
    mount_points: [
      { source_volume: 'data', container_path: '/data', read_only: true },
    ],
  },
  sidecars: {
    init: {
      essential: false,
      command: ['do-initialize', '/data'],
      mount_points: [
        { source_volume: 'data', container_path: '/data' },
      ],
    },
  },
```

In this case, `hako oneshot` should load the init sidecar by default for
convenience.